requirements.txt
# Core Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database & Caching
psycopg2-binary==2.9.9
redis==5.0.1

# Async HTTP & Processing
aiohttp==3.9.1
requests==2.31.0

# Configuration
python-dotenv==1.0.0
